Fix scrolled down bug in page transition
When the screen width is >= laptop range, there are some page
transitions that will fade the new page in a scrolled down state.
This is not the intended effect and is bad for ux. Additionaly, SignUp
content jumps down in this range because it is trying to vertically
center itself to the incoming content. This is jarring. So we have
removed the vertically centered styling from SignUp.